+++ lib/src/tuple8.dart
@@ -1,21 +1,28 @@
 /// # Tuple data structure
-class Tuple<T, S> {
+class Tuple<F, S> {
   const Tuple(this._first, this._second)
       : assert(
-          T != S,
+          F != S,
           'In this Tuple implementation, the types need to be different.',
         );
 
-  final T _first;
+  final F _first;
   final S _second;
 
   E? get<E>() {
-    if (E == T) return _first as E;
+    if (E == F) return _first as E;
     if (E == S) return _second as E;
     return null;
   }
 
-  Tuple<T, S> copyWith({T? first, S? second}) {
+  Tuple<F, S> copyWith({F? first, S? second}) {
     return Tuple(first ?? _first, second ?? _second);
   }
+
+  List<dynamic> toList({bool growable = false}) {
+    return List.from([_first, _second], growable: growable);
+  }
+
+  @override
+  String toString() => '($_first, $_second)';
 }